agentic-ai-strands-agents-and-mcp-retail-cpg/requirements.txt
strands-agents
strands-agents-tools
mcp
boto3
botocore
requests-aws4auth
pathlib
pillow
